% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_barlab_count.R
\name{geom_barlab_count}
\alias{geom_barlab_count}
\title{geom_text_count}
\usage{
geom_barlab_count(
  vjust = -0.15,
  position = ggplot2::position_dodge2(width = 0.9, preserve = "single"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Add label which is the count.  Acts like geom_bar() but text label at the position of the top of the bar.
}
